DL Sizing
DL Elements now resize dependent on the number of connection sockets on each side of the DL Element.
