+++ .gitignore
@@ -1,0 +1,5 @@
+
+*.pyc
+*.json
+*.pyc
+*.pyc
